Move custom file info file with same name
Best practise and also caused issues since dataclasses is the
module name from the standard libary
